Refactor: this includes: - Pull all section component to Section dir - Wrap SectionShadow to control card section
