+++ src/components/Section/SectionAction/ActionExecute.tsx
@@ -1,16 +1,21 @@
 import { Spacer } from "@nextui-org/react";
 import { observer } from "mobx-react-lite";
-import { ActionArchiveCards, cardStore } from "../../store/cardStore";
-import { ActionCard } from "../Card";
-import { Heading, HeadingDivider } from "../Heading";
+import { useEffect } from "react";
+import { ActionExecutedCards, cardStore } from "../../../store/cardStore";
+import { ActionCard } from "../../Card";
+import { Heading, HeadingDivider } from "../../Heading";
 
-const title = "已封存行動";
-const subtitle = "已封存";
+const title = "已執行行動";
+const subtitle = "已執行";
 
-export const ActionArchive = observer(() => {
+export const ActionExecute = observer(() => {
   const actionTodoCards = cardStore.getFilteredCardsWith(
-    new ActionArchiveCards(),
+    new ActionExecutedCards(),
   );
+
+  useEffect(() => {
+    cardStore.getArchivedCards();
+  }, [cardStore.archivedCards.length]);
 
   return (
     <>
