+++ src/components/Section/SectionIdea/IdeaAll.tsx
@@ -1,7 +1,7 @@
 import { observer } from "mobx-react-lite";
-import { IdeaAllCards, cardStore } from "../../store/cardStore";
-import { IdeaCard } from "../Card";
-import { Heading, HeadingDivider } from "../Heading";
+import { IdeaAllCards, cardStore } from "../../../store/cardStore";
+import { IdeaCard } from "../../Card";
+import { Heading, HeadingDivider } from "../../Heading";
 
 const title = "所有靈感";
 const subtitle = "靈感";
