+++ src/components/Section/SectionIdea/index.tsx
