+++ src/components/Section/SectionToday/TodayTodo.tsx
@@ -1,8 +1,9 @@
 import { Spacer } from "@nextui-org/react";
 import { observer } from "mobx-react-lite";
-import { TodoTodayCards, cardStore } from "../../store/cardStore";
-import { TodoCard } from "../Card";
-import EmptyCard from "../Card/EmptyCard";
+import { TodoTodayCards, cardStore } from "../../../store/cardStore";
+import { TodoCard } from "../../Card";
+import EmptyCard from "../../Card/EmptyCard";
+import SectionShadow from "../SectionShadow";
 
 const placeholder = "享受你的個人時光吧！";
 
@@ -20,12 +21,11 @@ export const TodayTodo = observer(() => {
 
   return (
     <section className="flex w-full flex-col items-center">
-      <div className="mt-5 flex w-full flex-col gap-3">
+      <SectionShadow className="h-[calc(100svh-313px)]">
         {todoCards.map((card) => (
           <TodoCard key={card.id} card={card} />
         ))}
-      </div>
-      <Spacer y={10} />
+      </SectionShadow>
     </section>
   );
 });
