+++ src/pages/TodayPage.tsx
@@ -2,8 +2,8 @@ import { Tab, Tabs } from "@nextui-org/react";
 import { observer } from "mobx-react-lite";
 import { Key, useEffect, useState } from "react";
 import { IdeaInput } from "../components/Input";
-import { TodayIdea, TodayTodo } from "../components/SectionToday";
-import CountBadge from "../components/SectionToday/CountBadge";
+import { TodayIdea, TodayTodo } from "../components/Section/SectionToday";
+import CountBadge from "../components/Section/SectionToday/CountBadge";
 import { authStore } from "../store/authStore";
 import { IdeaTodayCards, TodoTodayCards, cardStore } from "../store/cardStore";
 import { initTutorial, todaySteps } from "../utils/tutorial";
